requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.10.0
numpy>=1.24.0
python-pptx>=0.6.21
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
